tts: File Not Found verbose (#50)

+++ src/tts.py
@@ -8,6 +8,7 @@ import numpy as np
 import re
 import pandas as pd
 import sys
+from pathlib import Path
 import json
 import subprocess
 
@@ -88,13 +89,30 @@ class Synthesizer:
                     self._synthesize_line(phrases[i], voiceline_files[i])
             self.merge_audio_files(voiceline_files, final_voiceline_file)
 
-        # generate .lip file
-        self.run_command(f'{self.xvasynth_path}/resources/app/plugins/lip_fuz/FaceFXWrapper.exe "Skyrim" "USEnglish" "{self.xvasynth_path}/resources/app/plugins/lip_fuz/FonixData.cdf" "{final_voiceline_file}" "{final_voiceline_file.replace(".wav", "_r.wav")}" "{final_voiceline_file.replace(".wav", ".lip")}" "{voiceline}"')
+        if not os.path.exists(final_voiceline_file):
+            logging.error(f'xVASynth failed to generate voiceline at: {Path(final_voiceline_file)}')
+            raise FileNotFoundError()
+
+        # check if FonixData.cdf file is besides FaceFXWrapper.exe
+        cdf_path = f'{self.xvasynth_path}/resources/app/plugins/lip_fuz/FonixData.cdf'
+        if not os.path.exists(Path(cdf_path)):
+            logging.error(f'Could not find FonixData.cdf in "{Path(cdf_path).parent}" required by FaceFXWrapper. Look for the Lip Fuz plugin of xVASynth.')
+            raise FileNotFoundError()
+
+        # generate .lip file from the .wav file with FaceFXWrapper
+        face_wrapper_executable = f'{self.xvasynth_path}/resources/app/plugins/lip_fuz/FaceFXWrapper.exe';
+        if os.path.exists(face_wrapper_executable):
+            # Run FaceFXWrapper.exe
+            self.run_command(f'{face_wrapper_executable} "Skyrim" "USEnglish" "{self.xvasynth_path}/resources/app/plugins/lip_fuz/FonixData.cdf" "{final_voiceline_file}" "{final_voiceline_file.replace(".wav", "_r.wav")}" "{final_voiceline_file.replace(".wav", ".lip")}" "{voiceline}"')
+        else:
+            logging.error(f'Could not find FaceFXWrapper.exe in "{Path(face_wrapper_executable).parent}" with which to create a Lip Sync file, download it from: https://github.com/Nukem9/FaceFXWrapper/releases')
+            raise FileNotFoundError()
 
         # remove file created by FaceFXWrapper
         if os.path.exists(final_voiceline_file.replace(".wav", "_r.wav")):
             os.remove(final_voiceline_file.replace(".wav", "_r.wav"))
 
+        # if Debug Mode is on, play the audio file
         if (self.debug_mode == '1') & (self.play_audio_from_script == '1'):
             winsound.PlaySound(final_voiceline_file, winsound.SND_FILENAME)
